plagiarism_checker_service/flask_app/requirements.txt
Flask
openai
pymilvus
PyPDF2
python-docx
uuid
sentence_transformers